Movements updated V1.0
TX Update:
Had to update yaw, pitch, and roll to get accurate 1500 values to stabilize the drone.

RX update:
Will vary the throttle. It is needed because we have an xbox related joystick and it doesn't stick in one place. The only optimal values needed to for stabilizing the drone is min, middle and max throttle to keep the values stable hence the drone stable as well.