Addition of SD Card
Added sd card for data purpose to better tune PID. However, issue was that nRF24L01 didn't work synchornously with the SD card. It is either one or the other due to multiple SPI connection working together.

This will need to be explored in greater detail or use another smaller arduino which only works with SD card to get data.